prediction_service/web/core/requirements/local.txt

-r ./base.txt

# For SQL debug
django-silk>=4.1.0

# Test coverage
coverage==5.4

# Vulture finds unused code in Python programs.
vulture==2.3

# Debug Toolbar
django-debug-toolbar==3.2
django-debug-toolbar-request-history==0.1.3
